Make position / velocity floats

+++ src/game/gameobject.hpp
@@ -5,8 +5,8 @@ class GameObject {
     void set_pos(int x_pos, int y_pos);
 
   protected:
-    int x, y;
-    int vx, vy;
+    float x, y;
+    float vx, vy;
 };
 
 } // namespace game
